Time: 4 ms (29.79%), Space: 9 MB (98.11%) - LeetHub

+++ 0377-combination-sum-iv/0377-combination-sum-iv.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int combinationSum4(vector<int>& nums, int target) {   // 排列 vs. 518. 組合
-        vector<unsigned long long> dp(target + 1);
+        vector<unsigned int> dp(target + 1);
         dp[0] = 1;
         for (int i = 1; i <= target; i++) {
             for (int num : nums) {
